Issue #2 - Moved weapon_type functions to new file.
